+++ src/main/java/org/jabref/logic/msbib/MSBibEntry.java
@@ -4,9 +4,12 @@ import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+import org.jabref.model.entry.Author;
+import org.jabref.model.entry.AuthorList;
 import org.jabref.model.strings.StringUtil;
 
 import org.w3c.dom.Document;
@@ -26,21 +29,21 @@ class MSBibEntry {
     // MSBib fields and values
     public Map<String, String> fields = new HashMap<>();
 
-    public List<PersonName> authors;
-    public List<PersonName> bookAuthors;
-    public List<PersonName> editors;
-    public List<PersonName> translators;
-    public List<PersonName> producerNames;
-    public List<PersonName> composers;
-    public List<PersonName> conductors;
-    public List<PersonName> performers;
-    public List<PersonName> writers;
-    public List<PersonName> directors;
-    public List<PersonName> compilers;
-    public List<PersonName> interviewers;
-    public List<PersonName> interviewees;
-    public List<PersonName> inventors;
-    public List<PersonName> counsels;
+    public List<MsBibAuthor> authors;
+    public List<MsBibAuthor> bookAuthors;
+    public List<MsBibAuthor> editors;
+    public List<MsBibAuthor> translators;
+    public List<MsBibAuthor> producerNames;
+    public List<MsBibAuthor> composers;
+    public List<MsBibAuthor> conductors;
+    public List<MsBibAuthor> performers;
+    public List<MsBibAuthor> writers;
+    public List<MsBibAuthor> directors;
+    public List<MsBibAuthor> compilers;
+    public List<MsBibAuthor> interviewers;
+    public List<MsBibAuthor> interviewees;
+    public List<MsBibAuthor> inventors;
+    public List<MsBibAuthor> counsels;
 
     public PageNumbers pages;
 
@@ -203,8 +206,8 @@ class MSBibEntry {
         counsels = getSpecificAuthors("Counsel", authorsElem);
     }
 
-    private List<PersonName> getSpecificAuthors(String type, Element authors) {
-        List<PersonName> result = null;
+    private List<MsBibAuthor> getSpecificAuthors(String type, Element authors) {
+        List<MsBibAuthor> result = null;
         NodeList nodeLst = authors.getElementsByTagNameNS("*", type);
         if (nodeLst.getLength() <= 0) {
             return result;
@@ -223,17 +226,25 @@ class MSBibEntry {
             NodeList firstName = ((Element) person.item(i)).getElementsByTagNameNS("*", "First");
             NodeList lastName = ((Element) person.item(i)).getElementsByTagNameNS("*", "Last");
             NodeList middleName = ((Element) person.item(i)).getElementsByTagNameNS("*", "Middle");
-            PersonName name = new PersonName();
+
+            StringBuilder sb = new StringBuilder();
+
             if (firstName.getLength() > 0) {
-                name.setFirstname(firstName.item(0).getTextContent());
+                sb.append(firstName.item(0).getTextContent());
+                sb.append(" ");
             }
             if (middleName.getLength() > 0) {
-                name.setMiddlename(middleName.item(0).getTextContent());
+                sb.append(middleName.item(0).getTextContent());
+                sb.append(" ");
             }
             if (lastName.getLength() > 0) {
-                name.setSurname(lastName.item(0).getTextContent());
+                sb.append(lastName.item(0).getTextContent());
             }
-            result.add(name);
+
+            AuthorList authorList = AuthorList.parse(sb.toString());
+            for (Author author : authorList.getAuthors()) {
+                result.add(new MsBibAuthor(author));
+            }
         }
 
         return result;
@@ -286,6 +297,7 @@ class MSBibEntry {
         }
         addField(document, rootNode, "Year", year);
         addField(document, rootNode, "Month", month);
+        addField(document, rootNode, "Day", day);
 
         addField(document, rootNode, "JournalName", journalName);
         addField(document, rootNode, "PatentNumber", patentNumber);
@@ -316,21 +328,37 @@ class MSBibEntry {
         parent.appendChild(elem);
     }
 
-    private void addAuthor(Document document, Element allAuthors, String entryName, List<PersonName> authorsLst) {
+    //Add authors for export
+    private void addAuthor(Document document, Element allAuthors, String entryName, List<MsBibAuthor> authorsLst) {
         if (authorsLst == null) {
             return;
         }
         Element authorTop = document.createElementNS(MSBibDatabase.NAMESPACE, MSBibDatabase.PREFIX + entryName);
-        Element nameList = document.createElementNS(MSBibDatabase.NAMESPACE, MSBibDatabase.PREFIX + "NameList");
-        for (PersonName name : authorsLst) {
-            Element person = document.createElementNS(MSBibDatabase.NAMESPACE, MSBibDatabase.PREFIX + "Person");
-            addField(document, person, "Last", name.getSurname());
-            addField(document, person, "Middle", name.getMiddlename());
-            addField(document, person, "First", name.getFirstname());
-            nameList.appendChild(person);
+
+        Optional<MsBibAuthor> personName = authorsLst.stream().filter(MsBibAuthor::isCorporate)
+                .findFirst();
+        if (personName.isPresent()) {
+            MsBibAuthor person = personName.get();
+
+            Element corporate = document.createElementNS(MSBibDatabase.NAMESPACE,
+                    MSBibDatabase.PREFIX + "Corporate");
+            corporate.setTextContent(person.getFirstLast());
+            authorTop.appendChild(corporate);
+        } else {
+
+            Element nameList = document.createElementNS(MSBibDatabase.NAMESPACE, MSBibDatabase.PREFIX + "NameList");
+            for (MsBibAuthor name : authorsLst) {
+                Element person = document.createElementNS(MSBibDatabase.NAMESPACE, MSBibDatabase.PREFIX + "Person");
+                addField(document, person, "Last", name.getLastName());
+                addField(document, person, "Middle", name.getMiddleName());
+                addField(document, person, "First", name.getFirstName());
+                nameList.appendChild(person);
+
+            }
+            authorTop.appendChild(nameList);
         }
-        authorTop.appendChild(nameList);
         allAuthors.appendChild(authorTop);
+
     }
 
     private void addAddress(Document document, Element parent, String addressToSplit) {