+++ src/main/java/org/jabref/preferences/JabRefPreferences.java
@@ -182,11 +182,7 @@ public class JabRefPreferences {
     public static final String EDITOR_EMACS_KEYBINDINGS = "editorEMACSkeyBindings";
     public static final String EDITOR_EMACS_KEYBINDINGS_REBIND_CA = "editorEMACSkeyBindingsRebindCA";
     public static final String EDITOR_EMACS_KEYBINDINGS_REBIND_CF = "editorEMACSkeyBindingsRebindCF";
-    public static final String GROUP_EXPAND_TREE = "groupExpandTree";
-    public static final String GROUP_SHOW_DYNAMIC = "groupShowDynamic";
-    public static final String GROUP_SHOW_ICONS = "groupShowIcons";
     public static final String GROUPS_DEFAULT_FIELD = "groupsDefaultField";
-    public static final String GROUP_SHOW_OVERLAPPING = "groupShowOverlapping";
     public static final String GROUP_INVERT_SELECTIONS = "groupInvertSelections";
     public static final String GROUP_INTERSECT_SELECTIONS = "groupIntersectSelections";
     public static final String GROUP_FLOAT_SELECTIONS = "groupFloatSelections";
@@ -567,11 +563,7 @@ public class JabRefPreferences {
         defaults.put(GROUP_FLOAT_SELECTIONS, Boolean.TRUE);
         defaults.put(GROUP_INTERSECT_SELECTIONS, Boolean.TRUE);
         defaults.put(GROUP_INVERT_SELECTIONS, Boolean.FALSE);
-        defaults.put(GROUP_SHOW_OVERLAPPING, Boolean.FALSE);
         defaults.put(GROUPS_DEFAULT_FIELD, FieldName.KEYWORDS);
-        defaults.put(GROUP_SHOW_ICONS, Boolean.TRUE);
-        defaults.put(GROUP_SHOW_DYNAMIC, Boolean.TRUE);
-        defaults.put(GROUP_EXPAND_TREE, Boolean.TRUE);
         defaults.put(AUTO_ASSIGN_GROUP, Boolean.TRUE);
         defaults.put(KEYWORD_SEPARATOR, ", ");
         defaults.put(TOOLBAR_VISIBLE, Boolean.TRUE);