+++ CHANGELOG.md
@@ -20,6 +20,15 @@ We refer to [GitHub issues](https://github.com/JabRef/jabref/issues) by using `#
       - The background color of the hit counter signals whether the group contains all/any of the entries selected in the main table. 
       - Added a possibility to filter the groups panel [#1904](https://github.com/JabRef/jabref/issues/1904)
       - Removed edit mode.
+      - Removed the following commands in the right-click menu:
+        - Expand/collapse subtree
+        - Move up/down/left/right
+      - Remove option to "highlight overlapping groups"
+      - Moved the option to "Gray out non-hits" / "Hide non-hits" to the preferences
+      - Removed the following options from the group preferences:
+        - Show icons (icons can now be customized)
+        - Show dynamic groups in italics (dynamic groups are not treated specially now)
+        - Initially show groups tree expanded (always true now)
     - Redesigned about dialog.
     - Redesigned key bindings dialog.
     - Redesigned journal abbreviations dialog.
@@ -52,6 +61,9 @@ We refer to [GitHub issues](https://github.com/JabRef/jabref/issues) by using `#
 - We added an option to copy the title of BibTeX entries to the clipboard through `Edit -> Copy title` (implements [#210](https://github.com/koppor/jabref/issues/210))
 - Several scrollbars were added to the preference dialog which show up when content is too large [#2559](https://github.com/JabRef/jabref/issues/2559)
 - We fixed and improved the auto detection of the OpenOffice and LibreOffice connection
+- Entries with a single corporate author are now correclty exported to the corresponding `corporate` author field in MS-Office XML. [#1497](https://github.com/JabRef/jabref/issues/1497)
+- Improved author handling in MS-Office Import/Export
+- The `day` part of the biblatex `date` field is now exported to the corresponding `day` field in MS-Office XML. [#2691](https://github.com/JabRef/jabref/issues/2691)
 
 ### Fixed
  - We fixed an issue where authors with multiple surnames were not presented correctly in the main table. [#2534](https://github.com/JabRef/jabref/issues/2534)
@@ -66,7 +78,6 @@ We refer to [GitHub issues](https://github.com/JabRef/jabref/issues) by using `#
  - Fixed the synchronization logic of keywords and special fields and vice versa [#2580](https://github.com/JabRef/jabref/issues/2580)
  - We fixed an exception that prevented JabRef from starting in rare cases [bug report in the forum](http://discourse.jabref.org/t/jabref-not-opening/476).
  - We fixed an unhandled exception when saving an entry containing unbalanced braces [#2571](https://github.com/JabRef/jabref/issues/2571) 
- 
  - Fixed a display issue when removing a group with a long name [#1407](https://github.com/JabRef/jabref/issues/1407)
  - The group selection is no longer lost when switching tabs [#1104](https://github.com/JabRef/jabref/issues/1104)
  - We fixed an issue where the "find unlinked files" functionality threw an error when only one PDF was imported but not assigned to an entry [#2577](https://github.com/JabRef/jabref/issues/2577)
@@ -81,6 +92,7 @@ We refer to [GitHub issues](https://github.com/JabRef/jabref/issues) by using `#
  - We fixed an issue where the file folder could not be changed when running `Get fulltext` in the `general`-tab. [#2572](https://github.com/JabRef/jabref/issues/2572)
  - Newly created libraries no longer have the executable bit set under POSIX/Linux systems. The file permissions are now set to `664 (-rw-rw-r--)`. [#2635](https://github.com/JabRef/jabref/issues/#2635)
  - OpenOffice text formatting now handles nested tags properly [#2483](https://github.com/JabRef/jabref/issues/#2483)
+ - We fixed an issue where the dialog for selecting the main file directory in the preferences opened the the wrong folder
  
 ### Removed